# public/openapi.yaml
openapi: 3.1.0
info:
  title: mindlog · id — API
  version: 0.2.2
  description: |
    Carte d'identité en ligne (1 page + API + MCP).

    ## Authentification
    Deux mécanismes, résolus par `currentIdentity()` :
    1. **Cookie de session** `mindlog_session` (HttpOnly, Secure, SameSite=Lax) —
       posé après login passkey ou échange clé→session. Préféré pour le navigateur.
       Pour les méthodes non sûres (POST/PUT/PATCH/DELETE), l'en-tête `Origin` est
       vérifié quand il est présent (protection CSRF).
    2. **Clé d'accès** — secret de 24 octets (base64url). Transmis soit par l'en-tête
       `X-Access-Key`, soit par le paramètre de requête `?key=...`. Utilisé par les
       liens privés `/k/{clé}`, le serveur MCP et le flux SSE.

    Les routes « contacts » exigent une relation **réciproque** entre les deux handles.
    Les routes « publiques » n'exigent rien (un viewer authentifié élargit la réponse).

    ## Sécurité transverse
    - **Rate-limiting** (par IP) : `429 Too Many Requests` + `Retry-After`.
      Backstop global de 300 req/min sur `/api/*` ; limites strictes sur
      `POST /api/identities` (10/h), `POST /api/recover` (5/15 min),
      `POST /api/passkeys/auth/*` (20/15 min), `GET /api/search` (60/min).
    - **Taille du corps** : `413` au-delà de 256 Ko (JSON) ou 55 Mo (multipart).
    - **CSRF** : pour une mutation authentifiée par cookie, l'`Origin` (ou à
      défaut le `Referer`) doit correspondre à l'origine de l'app.
...
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mindlog_session
    accessKeyHeader:
      type: apiKey
      in: header
      name: X-Access-Key
    accessKeyQuery:
      type: apiKey
      in: query
      name: key
    oauth2:
      type: oauth2
      description: OAuth 2.1 (Authorization Code + PKCE S256, DCR). mindlog est AS + RS.
      flows:
        authorizationCode:
          authorizationUrl: https://id.mindlog.today/oauth/authorize
          tokenUrl: https://id.mindlog.today/oauth/token
          refreshUrl: https://id.mindlog.today/oauth/token
          scopes:
            "mindlog:identity": Gérer la carte d'identité du compte connecté
...
security:
  - sessionCookie: []
  - accessKeyHeader: []
  - accessKeyQuery: []
paths:
...
  /api/relations:
    post:
      tags: [Relations]
      summary: Ajouter une relation (par handle)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [handle]
              properties:
                handle: { type: string }
                type: { type: string, enum: [amis, pro, autre] }
      responses:
        '201': { description: Relation créée }
        '400': { description: handle requis / soi-même }
        '401': { description: Non authentifié }
        '404': { description: Handle inconnu }
  /api/relations/{handle}:
    delete:
      tags: [Relations]
      summary: Supprimer une relation
      parameters:
        - { name: handle, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Supprimée }
        '401': { description: Non authentifié }
        '404': { description: Introuvable }
...
  /api/messages/{handle}:
    get:
      tags: [Messagerie (E2E)]
      summary: Lire la conversation avec un contact
      parameters:
        - { name: handle, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Messages chiffrés + pubkey du pair }
        '403': { description: Réservé aux contacts (relation réciproque) }
    post:
      tags: [Messagerie (E2E)]
      summary: Envoyer un message chiffré (iv + ciphertext)
      parameters:
        - { name: handle, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                iv: { type: string }
                ciphertext: { type: string }
      responses:
        '201': { description: Envoyé }
        '400': { description: Erreur de stockage }
        '403': { description: Réservé aux contacts }
  /api/messages/{handle}/ack:
    post:
      tags: [Messagerie (E2E)]
      summary: Marquer les messages comme lus/non lus
      parameters:
        - { name: handle, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { read: { type: boolean } } }
      responses:
        '200': { description: OK }
        '403': { description: Réservé aux contacts }
  /api/messages/{handle}/{id}:
    delete:
      tags: [Messagerie (E2E)]
      summary: Supprimer un de ses messages
      parameters:
        - { name: handle, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Supprimé }
        '403': { description: Réservé aux contacts }
        '404': { description: Introuvable }
  /api/messages/{handle}/react:
    post:
      tags: [Messagerie (E2E)]
      summary: Ajouter/retirer une réaction emoji
      parameters:
        - { name: handle, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                messageId: { type: integer }
                emoji: { type: string }
      responses:
        '200': { description: OK }
        '400': { description: Invalide }
        '403': { description: Réservé aux contacts }
  /api/signal/{handle}:
    post:
      tags: [Messagerie (E2E)]
      summary: Relayer un signal WebRTC chiffré (appel P2P)
      description: >
        Relais éphémère de la signalisation WebRTC (offer/answer/ICE/hangup) vers
        un contact, poussé en temps réel via SSE (événement « signal »). Le serveur
        ne stocke rien et ne peut pas lire le contenu (chiffré de bout en bout).
        Le média audio/vidéo, lui, circule en P2P direct entre navigateurs.
      parameters:
        - { name: handle, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                iv: { type: string }
                ciphertext: { type: string, description: "Payload SDP/ICE chiffré (max 16000 car.)" }
      responses:
        '200': { description: Relayé au pair }
        '400': { description: Signal invalide }
        '403': { description: Réservé aux contacts }
...
  /api/notifications/read:
    post:
      tags: [Notifications & Temps réel]
      summary: Marquer toutes les notifications comme lues
      responses:
        '200': { description: OK }
        '401': { description: Non authentifié }
...